+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-31 21:53:12
-// Commit: feat(ui): implement responsive ui view component
+// Updated at: 2026-06-01 21:37:16
+// Commit: fix(ui): prevent double tap duplicate action in ui
 
 class uiComponentState {
-  final String commitRef = '1661';
-  final String lastMessage = 'feat(ui): implement responsive ui view component';
-  final DateTime timestamp = DateTime.parse('2026-05-31T21:53:12+05:00');
+  final String commitRef = '1677';
+  final String lastMessage = 'fix(ui): prevent double tap duplicate action in ui';
+  final DateTime timestamp = DateTime.parse('2026-06-01T21:37:16+05:00');
 }